Modules/MessagePassing/requirements.txt
google-cloud-firestore
boto3
functions-framework
